#11 refactor: split large files into logical submodules
Split postgres.rs (783 lines) into postgres/ directory:
- mod.rs: Main generator entry point
- context.rs: Generation context with precomputed values
- crud.rs: CREATE, READ, UPDATE, DELETE, LIST methods
- query.rs: Type-safe query filtering method
- relations.rs: belongs_to and has_many methods
- projections.rs: Optimized projection SELECT methods
- soft_delete.rs: Soft delete support methods
- helpers.rs: SQL building helper functions

Split entity.rs (636 lines) into entity/ directory:
- mod.rs: EntityDef struct and implementation
- attrs.rs: EntityAttrs (darling parsing struct)
- projection.rs: ProjectionDef and parsing
